{"schema":"decodex/commit/1","summary":"Split StateStore runtime modules","authority":"XY-1115"}
