+++ apps/decodex/src/state/store.rs
@@ -1,14 +1,4 @@
 use crate::autonomy_proposal::AutonomyProposalChallengeInput;
-use project_run_recovery::{
-	ProjectRunListingMode, project_lease_run_ids,
-	project_run_recovery_candidate_counts_as_project_run, project_run_recovery_candidates,
-	project_run_status_from_recovery_candidate,
-};
-
-const TERMINAL_THREAD_ARCHIVE_EVENT_TYPES: [&str; 2] =
-	["thread/archive", "thread/archive/discarded"];
-const DISCARDED_POST_ARCHIVE_PROTOCOL_EVENT_TYPE: &str = "protocol/post_archive_event/discarded";
-const REVIEW_CHECKPOINT_PROMPT_VERSION: &str = "decodex-review-checkpoint/2";
 
 /// Input fields for recording a project-scoped external connector backoff.
 pub(crate) struct ConnectorBackoffInput<'a> {
@@ -1016,756 +1006,6 @@ impl StateStore {
 		state.remember_run_project(project_id, issue_id, Some(run_id));
 
 		self.persist_runtime_state_locked(&state)
-	}
-
-	/// Insert or update a run attempt record.
-	pub fn record_run_attempt(
-		&self,
-		run_id: &str,
-		issue_id: &str,
-		attempt_number: i64,
-		status: &str,
-	) -> Result<()> {
-		let now = timestamp_parts();
-		let mut state = self.lock_without_refresh()?;
-		let project_id = state.project_id_for_run(issue_id, run_id);
-
-		match state.run_attempts.get_mut(run_id) {
-			Some(existing) => {
-				let retained_project_id =
-					(existing.issue_id == issue_id).then(|| existing.project_id.clone()).flatten();
-
-				existing.issue_id = issue_id.to_owned();
-				existing.project_id = project_id.or(retained_project_id);
-				existing.attempt_number = attempt_number;
-				existing.status = status.to_owned();
-				existing.updated_at = now.text.clone();
-				existing.updated_at_unix = now.unix;
-			},
-			None => {
-				state.run_attempts.insert(
-					run_id.to_owned(),
-					RunAttemptRecord {
-						run_id: run_id.to_owned(),
-						project_id,
-						issue_id: issue_id.to_owned(),
-						attempt_number,
-						status: status.to_owned(),
-						thread_id: None,
-						turn_id: None,
-						updated_at: now.text,
-						updated_at_unix: now.unix,
-					},
-				);
-			},
-		}
-
-		let attempt = state
-			.run_attempts
-			.get(run_id)
-			.ok_or_else(|| eyre::eyre!("Run attempt `{run_id}` was not recorded."))?
-			.clone();
-
-		self.upsert_run_attempt_locked(&attempt)
-	}
-
-	/// Compute the next attempt number for one issue.
-	pub fn next_attempt_number(&self, issue_id: &str) -> Result<i64> {
-		let state = self.lock()?;
-		let next_attempt = state
-			.run_attempts
-			.values()
-			.filter(|attempt| attempt.issue_id == issue_id)
-			.map(|attempt| attempt.attempt_number)
-			.max()
-			.unwrap_or(0)
-			+ 1;
-
-		Ok(next_attempt)
-	}
-
-	/// Count attempts that consume the retry budget for one issue.
-	pub fn retry_budget_attempt_count(&self, issue_id: &str) -> Result<i64> {
-		if let Some(sqlite) = self.sqlite.as_ref() {
-			let sqlite =
-				sqlite.lock().map_err(|_| eyre::eyre!("StateStore SQLite mutex is poisoned."))?;
-
-			return sqlite.retry_budget_attempt_count(issue_id);
-		}
-
-		let state = self.lock_without_refresh()?;
-		let retry_budget_attempts = state
-			.run_attempts
-			.values()
-			.filter(|attempt| {
-				attempt.issue_id == issue_id
-					&& matches!(
-						attempt.status.as_str(),
-						"failed" | "interrupted" | "terminal_guarded"
-					)
-			})
-			.count() as i64;
-
-		Ok(retry_budget_attempts)
-	}
-
-	/// Return whether a later attempt for one issue consumed retry budget.
-	pub fn issue_has_retry_budget_attempt_after(
-		&self,
-		issue_id: &str,
-		attempt_number: i64,
-	) -> Result<bool> {
-		if let Some(sqlite) = self.sqlite.as_ref() {
-			let sqlite =
-				sqlite.lock().map_err(|_| eyre::eyre!("StateStore SQLite mutex is poisoned."))?;
-
-			return sqlite.issue_has_retry_budget_attempt_after(issue_id, attempt_number);
-		}
-
-		let state = self.lock_without_refresh()?;
-
-		Ok(state.run_attempts.values().any(|attempt| {
-			attempt.issue_id == issue_id
-				&& attempt.attempt_number > attempt_number
-				&& matches!(attempt.status.as_str(), "failed" | "interrupted" | "terminal_guarded")
-		}))
-	}
-
-	/// Attach the active thread identifier to a run attempt.
-	pub fn update_run_thread(&self, run_id: &str, thread_id: &str) -> Result<()> {
-		let now = timestamp_parts();
-		let mut state = self.lock_without_refresh()?;
-
-		if let Some(attempt) = state.run_attempts.get_mut(run_id) {
-			attempt.thread_id = Some(thread_id.to_owned());
-			attempt.updated_at = now.text;
-			attempt.updated_at_unix = now.unix;
-
-			let attempt = attempt.clone();
-
-			return self.upsert_run_attempt_locked(&attempt);
-		}
-
-		Ok(())
-	}
-
-	/// Attach the active turn identifier to a run attempt.
-	pub fn update_run_turn(&self, run_id: &str, turn_id: &str) -> Result<()> {
-		let now = timestamp_parts();
-		let mut state = self.lock_without_refresh()?;
-
-		if let Some(attempt) = state.run_attempts.get_mut(run_id) {
-			attempt.turn_id = Some(turn_id.to_owned());
-			attempt.updated_at = now.text;
-			attempt.updated_at_unix = now.unix;
-
-			let attempt = attempt.clone();
-
-			return self.upsert_run_attempt_locked(&attempt);
-		}
-
-		Ok(())
-	}
-
-	/// Update the status for one run attempt.
-	pub fn update_run_status(&self, run_id: &str, status: &str) -> Result<()> {
-		let now = timestamp_parts();
-		let mut state = self.lock_without_refresh()?;
-
-		if let Some(attempt) = state.run_attempts.get_mut(run_id) {
-			attempt.status = status.to_owned();
-			attempt.updated_at = now.text;
-			attempt.updated_at_unix = now.unix;
-
-			let attempt = attempt.clone();
-
-			return self.upsert_run_attempt_locked(&attempt);
-		}
-
-		Ok(())
-	}
-
-	/// Mark all running run attempts for one issue as succeeded.
-	pub fn succeed_running_run_attempts_for_issue(&self, issue_id: &str) -> Result<usize> {
-		let now = timestamp_parts();
-		let mut state = self.lock()?;
-		let mut updated_count = 0;
-
-		for attempt in state
-			.run_attempts
-			.values_mut()
-			.filter(|attempt| attempt.issue_id == issue_id)
-			.filter(|attempt| running_run_attempt_status(&attempt.status))
-		{
-			attempt.status = "succeeded".to_owned();
-			attempt.updated_at = now.text.clone();
-			attempt.updated_at_unix = now.unix;
-			updated_count += 1;
-		}
-
-		if updated_count > 0 {
-			self.persist_runtime_state_locked(&state)?;
-		}
-
-		Ok(updated_count)
-	}
-
-	/// Read one run attempt.
-	pub fn run_attempt(&self, run_id: &str) -> Result<Option<RunAttempt>> {
-		let state = self.lock()?;
-
-		Ok(state.run_attempts.get(run_id).map(RunAttemptRecord::as_public))
-	}
-
-	/// Read one run attempt by issue and attempt number.
-	pub fn run_attempt_for_issue_attempt(
-		&self,
-		issue_id: &str,
-		attempt_number: i64,
-	) -> Result<Option<RunAttempt>> {
-		if let Some(sqlite) = &self.sqlite {
-			let sqlite = sqlite.lock().map_err(|_| eyre::eyre!("State store lock poisoned."))?;
-
-			return sqlite
-				.run_attempt_for_issue_attempt(issue_id, attempt_number)
-				.map(|attempt| attempt.map(|attempt| attempt.as_public()));
-		}
-
-		let state = self.lock()?;
-		let attempt = state
-			.run_attempts
-			.values()
-			.filter(|attempt| {
-				attempt.issue_id == issue_id && attempt.attempt_number == attempt_number
-			})
-			.max_by(|left, right| compare_attempt_records(left, right))
-			.map(RunAttemptRecord::as_public);
-
-		Ok(attempt)
-	}
-
-	/// Read the latest run attempt for one issue.
-	pub fn latest_run_attempt_for_issue(&self, issue_id: &str) -> Result<Option<RunAttempt>> {
-		if let Some(sqlite) = &self.sqlite {
-			let sqlite = sqlite.lock().map_err(|_| eyre::eyre!("State store lock poisoned."))?;
-
-			return sqlite
-				.latest_run_attempt_for_issue(issue_id)
-				.map(|attempt| attempt.map(|attempt| attempt.as_public()));
-		}
-
-		let state = self.lock()?;
-		let attempt = state
-			.run_attempts
-			.values()
-			.filter(|attempt| attempt.issue_id == issue_id)
-			.max_by(|left, right| compare_attempt_records(left, right))
-			.map(RunAttemptRecord::as_public);
-
-		Ok(attempt)
-	}
-
-	/// List all locally recorded run attempts for one issue.
-	pub fn list_run_attempts_for_issue(&self, issue_id: &str) -> Result<Vec<RunAttempt>> {
-		if let Some(sqlite) = &self.sqlite {
-			let sqlite = sqlite.lock().map_err(|_| eyre::eyre!("State store lock poisoned."))?;
-			let attempts = sqlite
-				.list_run_attempts_for_issue(issue_id)?
-				.into_iter()
-				.map(|attempt| attempt.as_public())
-				.collect();
-
-			return Ok(attempts);
-		}
-
-		let state = self.lock()?;
-		let mut attempts = state
-			.run_attempts
-			.values()
-			.filter(|attempt| attempt.issue_id == issue_id)
-			.map(RunAttemptRecord::as_public)
-			.collect::<Vec<_>>();
-
-		attempts.sort_by(|left, right| {
-			left.attempt_number()
-				.cmp(&right.attempt_number())
-				.then_with(|| left.run_id().cmp(right.run_id()))
-		});
-
-		Ok(attempts)
-	}
-
-	/// List all locally recorded run attempts for one registered project.
-	pub fn list_run_attempts_for_project(&self, project_id: &str) -> Result<Vec<RunAttempt>> {
-		if let Some(sqlite) = &self.sqlite {
-			let sqlite = sqlite.lock().map_err(|_| eyre::eyre!("State store lock poisoned."))?;
-			let attempts = sqlite
-				.list_run_attempts_for_project(project_id)?
-				.into_iter()
-				.map(|attempt| attempt.as_public())
-				.collect();
-
-			return Ok(attempts);
-		}
-
-		let state = self.lock()?;
-		let mut attempts = state
-			.run_attempts
-			.values()
-			.filter(|attempt| attempt.project_id.as_deref() == Some(project_id))
-			.map(RunAttemptRecord::as_public)
-			.collect::<Vec<_>>();
-
-		attempts.sort_by(|left, right| right.run_id().cmp(left.run_id()));
-
-		Ok(attempts)
-	}
-
-	/// Return whether one run already has a matching protocol event.
-	pub fn run_has_protocol_event(&self, run_id: &str, event_type: &str) -> Result<bool> {
-		if let Some(sqlite) = &self.sqlite {
-			let sqlite = sqlite.lock().map_err(|_| eyre::eyre!("State store lock poisoned."))?;
-
-			return sqlite.run_has_protocol_event(run_id, event_type);
-		}
-
-		let state = self.lock()?;
-
-		Ok(state
-			.events
-			.get(run_id)
-			.is_some_and(|events| events.iter().any(|event| event.event_type == event_type)))
-	}
-
-	/// List recent run attempts for one project, including lease and protocol summary fields.
-	pub fn list_recent_runs(
-		&self,
-		project_id: &str,
-		limit: usize,
-	) -> Result<Vec<ProjectRunStatus>> {
-		self.list_project_runs(project_id, limit).map(|(_active, recent)| recent)
-	}
-
-	/// List active and recent run attempts for one project from one durable snapshot.
-	pub(crate) fn list_project_runs(
-		&self,
-		project_id: &str,
-		base_recent_limit: usize,
-	) -> Result<(Vec<ProjectRunStatus>, Vec<ProjectRunStatus>)> {
-		self.list_project_runs_with_mode(
-			project_id,
-			base_recent_limit,
-			ProjectRunListingMode::AllowMarkerIdentityPersistence,
-		)
-	}
-
-	/// List active and recent project runs without persisting marker-derived identities.
-	pub(crate) fn list_project_runs_read_only(
-		&self,
-		project_id: &str,
-		base_recent_limit: usize,
-	) -> Result<(Vec<ProjectRunStatus>, Vec<ProjectRunStatus>)> {
-		self.list_project_runs_with_mode(
-			project_id,
-			base_recent_limit,
-			ProjectRunListingMode::ReadOnly,
-		)
-	}
-
-	fn list_project_runs_with_mode(
-		&self,
-		project_id: &str,
-		base_recent_limit: usize,
-		mode: ProjectRunListingMode,
-	) -> Result<(Vec<ProjectRunStatus>, Vec<ProjectRunStatus>)> {
-		let mut state = self.lock_without_refresh()?;
-
-		self.refresh_project_run_metadata_state_locked(&mut state, project_id)?;
-
-		if matches!(mode, ProjectRunListingMode::AllowMarkerIdentityPersistence) {
-			self.refresh_run_attempt_identities_from_worktree_markers_locked(
-				&mut state, project_id,
-			)?;
-		}
-
-		self.refresh_project_loop_evidence_state_locked(&mut state, project_id)?;
-
-		let lease_run_ids = project_lease_run_ids(&state, project_id, None);
-
-		self.refresh_run_activity_summaries_for_runs_locked(&mut state, &lease_run_ids)?;
-
-		let recovery_candidates = project_run_recovery_candidates(&state, project_id, None)?
-			.into_iter()
-			.filter(|candidate| {
-				project_run_recovery_candidate_counts_as_project_run(&state, candidate)
-			})
-			.collect::<Vec<_>>();
-		let recovery_run_ids = recovery_candidates
-			.iter()
-			.map(|candidate| candidate.run_id().to_owned())
-			.collect::<Vec<_>>();
-
-		self.refresh_run_activity_summaries_for_runs_locked(&mut state, &recovery_run_ids)?;
-		self.refresh_protocol_event_summaries_for_runs_locked(&mut state, &recovery_run_ids)?;
-
-		let mut runs = state
-			.run_attempts
-			.values()
-			.filter_map(|attempt| state.project_run_status(project_id, attempt))
-			.collect::<Vec<_>>();
-
-		runs.extend(
-			recovery_candidates.iter().filter_map(|candidate| {
-				project_run_status_from_recovery_candidate(&state, candidate)
-			}),
-		);
-		runs.sort_by(compare_project_run_status);
-
-		let leased_runs =
-			runs.iter().filter(|status| status.run_lease()).cloned().collect::<Vec<_>>();
-		let recent_limit = base_recent_limit.saturating_add(leased_runs.len());
-		let recent_run_ids =
-			runs.iter().take(recent_limit).map(|run| run.run_id().to_owned()).collect::<Vec<_>>();
-		let mut summary_run_ids =
-			leased_runs.iter().map(|run| run.run_id().to_owned()).collect::<Vec<_>>();
-
-		summary_run_ids.extend(recent_run_ids);
-		summary_run_ids.sort();
-		summary_run_ids.dedup();
-		self.refresh_protocol_event_summaries_for_runs_locked(&mut state, &summary_run_ids)?;
-		self.refresh_run_activity_summaries_for_runs_locked(&mut state, &summary_run_ids)?;
-
-		let summary_run_id_set = summary_run_ids.iter().cloned().collect::<HashSet<_>>();
-		let mut selected_runs = state
-			.run_attempts
-			.values()
-			.filter(|attempt| summary_run_id_set.contains(&attempt.run_id))
-			.filter_map(|attempt| state.project_run_status(project_id, attempt))
-			.collect::<Vec<_>>();
-
-		selected_runs.extend(
-			recovery_candidates
-				.iter()
-				.filter(|candidate| summary_run_id_set.contains(candidate.run_id()))
-				.filter_map(|candidate| {
-					project_run_status_from_recovery_candidate(&state, candidate)
-				}),
-		);
-		selected_runs.sort_by(compare_project_run_status);
-
-		let leased_runs =
-			selected_runs.iter().filter(|status| status.run_lease()).cloned().collect::<Vec<_>>();
-		let mut recent_runs = selected_runs;
-
-		recent_runs.truncate(recent_limit);
-
-		Ok((leased_runs, recent_runs))
-	}
-
-	/// List all locally recorded run attempts for one issue in one project.
-	pub(crate) fn list_project_issue_runs(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-	) -> Result<Vec<ProjectRunStatus>> {
-		let mut state = self.lock_without_refresh()?;
-
-		self.refresh_project_run_metadata_state_locked(&mut state, project_id)?;
-		self.refresh_run_attempt_identities_from_worktree_markers_locked(&mut state, project_id)?;
-		self.refresh_project_loop_evidence_state_locked(&mut state, project_id)?;
-
-		let lease_run_ids = project_lease_run_ids(&state, project_id, Some(issue_id));
-
-		self.refresh_run_activity_summaries_for_runs_locked(&mut state, &lease_run_ids)?;
-
-		let recovery_candidates =
-			project_run_recovery_candidates(&state, project_id, Some(issue_id))?;
-		let recovery_run_ids = recovery_candidates
-			.iter()
-			.map(|candidate| candidate.run_id().to_owned())
-			.collect::<Vec<_>>();
-		let run_ids = state
-			.run_attempts
-			.values()
-			.filter(|attempt| attempt.issue_id == issue_id)
-			.map(|attempt| attempt.run_id.clone())
-			.collect::<Vec<_>>();
-		let mut summary_run_ids = run_ids;
-
-		summary_run_ids.extend(recovery_run_ids);
-		summary_run_ids.sort();
-		summary_run_ids.dedup();
-		self.refresh_run_activity_summaries_for_runs_locked(&mut state, &summary_run_ids)?;
-		self.refresh_protocol_event_summaries_for_runs_locked(&mut state, &summary_run_ids)?;
-
-		let mut runs = state
-			.run_attempts
-			.values()
-			.filter(|attempt| attempt.issue_id == issue_id)
-			.filter_map(|attempt| state.project_run_status(project_id, attempt))
-			.collect::<Vec<_>>();
-
-		runs.extend(
-			recovery_candidates.iter().filter_map(|candidate| {
-				project_run_status_from_recovery_candidate(&state, candidate)
-			}),
-		);
-		runs.sort_by(compare_project_run_status);
-
-		Ok(runs)
-	}
-
-	/// List all leased run attempts for one project without applying the recent-run limit.
-	pub fn list_leased_runs(&self, project_id: &str) -> Result<Vec<ProjectRunStatus>> {
-		let mut state = self.lock_without_refresh()?;
-
-		self.refresh_project_run_metadata_state_locked(&mut state, project_id)?;
-
-		let mut runs = state
-			.run_attempts
-			.values()
-			.filter_map(|attempt| {
-				let status = state.project_run_status(project_id, attempt)?;
-
-				status.run_lease.then_some(status)
-			})
-			.collect::<Vec<_>>();
-		let mut run_ids = runs.iter().map(|run| run.run_id().to_owned()).collect::<Vec<_>>();
-
-		run_ids.sort();
-		run_ids.dedup();
-		self.refresh_protocol_event_summaries_for_runs_locked(&mut state, &run_ids)?;
-
-		runs = state
-			.run_attempts
-			.values()
-			.filter(|attempt| run_ids.contains(&attempt.run_id))
-			.filter_map(|attempt| {
-				let status = state.project_run_status(project_id, attempt)?;
-
-				status.run_lease.then_some(status)
-			})
-			.collect::<Vec<_>>();
-
-		runs.sort_by(compare_project_run_status);
-
-		Ok(runs)
-	}
-
-	/// Append one protocol event to the journal for a run.
-	pub fn append_event(
-		&self,
-		run_id: &str,
-		sequence_number: i64,
-		event_type: &str,
-		payload: &str,
-	) -> Result<()> {
-		let now = timestamp_parts();
-		let mut state = self.lock_without_refresh()?;
-		let event = ProtocolEventRecord {
-			sequence_number,
-			event_type: event_type.to_owned(),
-			payload_sha256: protocol_event_payload_sha256(payload),
-			created_at: now.text,
-			created_at_unix: now.unix,
-		};
-		let Some(mut event) =
-			self.prepare_protocol_event_for_append_locked(&mut state, run_id, event)?
-		else {
-			return Ok(());
-		};
-
-		loop {
-			let (insert_index, cached_existing) = {
-				let events = state.events.entry(run_id.to_owned()).or_default();
-
-				match events
-					.binary_search_by_key(&event.sequence_number, |event| event.sequence_number)
-				{
-					Ok(index) => (index, Some(events[index].clone())),
-					Err(index) => (index, None),
-				}
-			};
-
-			if let Some(existing) = cached_existing {
-				if self.handle_protocol_event_append_conflict_locked(
-					&mut state, run_id, &mut event, &existing,
-				)? {
-					continue;
-				}
-
-				return Ok(());
-			}
-
-			if !self.append_protocol_event_locked(run_id, &event)? {
-				let existing =
-					self.protocol_event_locked(run_id, event.sequence_number)?.ok_or_else(|| {
-						eyre::eyre!(
-							"Protocol event `{run_id}` sequence `{}` already exists in the runtime journal, but the existing row could not be read.",
-							event.sequence_number
-						)
-					})?;
-
-				if self.handle_protocol_event_append_conflict_locked(
-					&mut state, run_id, &mut event, &existing,
-				)? {
-					continue;
-				}
-
-				return Ok(());
-			}
-
-			self.record_inserted_protocol_event_locked(&mut state, run_id, insert_index, event)?;
-
-			return Ok(());
-		}
-	}
-
-	fn prepare_protocol_event_for_append_locked(
-		&self,
-		state: &mut StateData,
-		run_id: &str,
-		event: ProtocolEventRecord,
-	) -> Result<Option<ProtocolEventRecord>> {
-		if !self.protocol_event_should_be_discarded_after_archive_locked(state, run_id, &event)? {
-			return Ok(Some(event));
-		}
-		if self.protocol_event_replay_already_recorded_locked(state, run_id, &event)? {
-			self.refresh_protocol_event_summaries_for_runs_locked(state, &[run_id.to_owned()])?;
-
-			return Ok(None);
-		}
-
-		Ok(Some(discarded_post_archive_protocol_event_with_log(run_id, event)))
-	}
-
-	fn handle_protocol_event_append_conflict_locked(
-		&self,
-		state: &mut StateData,
-		run_id: &str,
-		event: &mut ProtocolEventRecord,
-		existing: &ProtocolEventRecord,
-	) -> Result<bool> {
-		if protocol_event_conflict_should_be_discarded_after_archive(existing, event) {
-			*event = discarded_post_archive_protocol_event_with_log(run_id, event.clone());
-
-			return Ok(true);
-		}
-		if protocol_event_is_discarded_post_archive_collision(existing, event) {
-			event.sequence_number =
-				next_discarded_post_archive_sequence_after_collision(event.sequence_number)?;
-
-			return Ok(true);
-		}
-
-		ensure_protocol_event_replay_matches(run_id, existing, event)?;
-
-		self.refresh_protocol_event_summaries_for_runs_locked(state, &[run_id.to_owned()])?;
-
-		Ok(false)
-	}
-
-	fn record_inserted_protocol_event_locked(
-		&self,
-		state: &mut StateData,
-		run_id: &str,
-		insert_index: usize,
-		event: ProtocolEventRecord,
-	) -> Result<()> {
-		let had_cached_summary = state.event_summaries.contains_key(run_id);
-		let inserted_event = event.clone();
-
-		state.events.entry(run_id.to_owned()).or_default().insert(insert_index, event);
-
-		if self.sqlite.is_some() && !had_cached_summary {
-			self.rebuild_protocol_event_summaries_for_runs_locked(state, &[run_id.to_owned()])?;
-		} else if self.sqlite.is_some() {
-			let summary = state.event_summaries.entry(run_id.to_owned()).or_default();
-
-			if summary.last_sequence_number.is_none_or(|last_sequence_number| {
-				inserted_event.sequence_number == last_sequence_number.saturating_add(1)
-			}) {
-				summary.record_event(&inserted_event);
-				let summary = summary.clone();
-
-				self.upsert_protocol_event_summary_locked(run_id, &summary)?;
-			} else {
-				self.rebuild_protocol_event_summaries_for_runs_locked(state, &[run_id.to_owned()])?;
-			}
-		} else if let Some(events) = state.events.get(run_id) {
-			let summary = protocol_event_summary_from_events(events);
-
-			state.event_summaries.insert(run_id.to_owned(), summary);
-		}
-
-		Ok(())
-	}
-
-	fn protocol_event_should_be_discarded_after_archive_locked(
-		&self,
-		state: &StateData,
-		run_id: &str,
-		event: &ProtocolEventRecord,
-	) -> Result<bool> {
-		if !protocol_event_can_be_discarded_after_archive(event) {
-			return Ok(false);
-		}
-
-		self.run_has_terminal_thread_archive_event_locked(state, run_id)
-	}
-
-	fn run_has_terminal_thread_archive_event_locked(
-		&self,
-		state: &StateData,
-		run_id: &str,
-	) -> Result<bool> {
-		if state.events.get(run_id).is_some_and(|events| {
-			events.iter().any(|event| protocol_event_is_terminal_thread_archive(&event.event_type))
-		}) {
-			return Ok(true);
-		}
-		if state.event_summaries.get(run_id).is_some_and(|summary| {
-			summary
-				.last_event_type
-				.as_deref()
-				.is_some_and(protocol_event_is_terminal_thread_archive)
-		}) {
-			return Ok(true);
-		}
-
-		let Some(sqlite) = self.sqlite.as_ref() else {
-			return Ok(false);
-		};
-		let sqlite =
-			sqlite.lock().map_err(|_| eyre::eyre!("StateStore SQLite mutex is poisoned."))?;
-
-		for event_type in TERMINAL_THREAD_ARCHIVE_EVENT_TYPES {
-			if sqlite.run_has_protocol_event(run_id, event_type)? {
-				return Ok(true);
-			}
-		}
-
-		Ok(false)
-	}
-
-	fn protocol_event_replay_already_recorded_locked(
-		&self,
-		state: &StateData,
-		run_id: &str,
-		event: &ProtocolEventRecord,
-	) -> Result<bool> {
-		if let Some(events) = state.events.get(run_id)
-			&& let Ok(index) =
-				events.binary_search_by_key(&event.sequence_number, |event| event.sequence_number)
-		{
-			return Ok(events[index].is_idempotent_replay_of(event));
-		}
-
-		let Some(existing) = self.protocol_event_locked(run_id, event.sequence_number)? else {
-			return Ok(false);
-		};
-
-		Ok(existing.is_idempotent_replay_of(event))
 	}
 
 	pub(crate) fn record_run_activity_summary(
@@ -3234,13 +2474,6 @@ impl StateStore {
 		Ok(records)
 	}
 
-	/// Count protocol journal records for one run.
-	pub fn event_count(&self, run_id: &str) -> Result<i64> {
-		let state = self.lock()?;
-
-		Ok(state.protocol_event_summary(run_id).event_count)
-	}
-
 	/// Read the latest recorded activity timestamp for one run as a Unix epoch.
 	pub fn last_run_activity_unix_epoch(&self, run_id: &str) -> Result<Option<i64>> {
 		let state = self.lock()?;
@@ -3253,614 +2486,6 @@ impl StateStore {
 			(None, Some(event_activity)) => Some(event_activity),
 			(None, None) => None,
 		})
-	}
-
-	/// Read the latest recorded protocol-event timestamp for one run as a Unix epoch.
-	pub fn last_protocol_activity_unix_epoch(&self, run_id: &str) -> Result<Option<i64>> {
-		let state = self.lock()?;
-
-		Ok(state.protocol_event_summary(run_id).last_event_at_unix)
-	}
-
-	/// Create or replace the worktree mapping for one issue.
-	pub fn upsert_worktree(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-		branch_name: &str,
-		worktree_path: &str,
-	) -> Result<()> {
-		let mut state = self.lock_without_refresh()?;
-		let now_unix = OffsetDateTime::now_utc().unix_timestamp();
-		let created_at_unix = state
-			.worktrees
-			.get(issue_id)
-			.and_then(|mapping| mapping.created_at_unix)
-			.or(Some(now_unix));
-		let mapping = WorktreeMappingRecord {
-			project_id: project_id.to_owned(),
-			issue_id: issue_id.to_owned(),
-			branch_name: branch_name.to_owned(),
-			worktree_path: PathBuf::from(worktree_path),
-			provenance_source: WORKTREE_PROVENANCE_RUNTIME_RECORDED.to_owned(),
-			created_at_unix,
-			updated_at_unix: Some(now_unix),
-		};
-
-		state.worktrees.insert(issue_id.to_owned(), mapping.clone());
-		state.remember_run_project(project_id, issue_id, None);
-
-		self.upsert_worktree_and_remember_run_project_locked(&mapping)
-	}
-
-	/// Create or refresh a worktree mapping reconstructed from retained local state.
-	pub(crate) fn upsert_recovered_worktree(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-		branch_name: &str,
-		worktree_path: &str,
-		observed_at_unix: Option<i64>,
-	) -> Result<()> {
-		let mut state = self.lock_without_refresh()?;
-		let existing = state.worktrees.get(issue_id);
-		let existing_provenance_source = existing.map(|mapping| mapping.provenance_source.as_str());
-		let provenance_source = match existing_provenance_source {
-			Some(WORKTREE_PROVENANCE_RUNTIME_RECORDED) => WORKTREE_PROVENANCE_RUNTIME_RECORDED,
-			Some(WORKTREE_PROVENANCE_RUNTIME_RECOVERED) => WORKTREE_PROVENANCE_RUNTIME_RECOVERED,
-			_ => WORKTREE_PROVENANCE_RUNTIME_RECOVERED,
-		}
-		.to_owned();
-		let existing_created_at_unix = existing.and_then(|mapping| mapping.created_at_unix);
-		let existing_updated_at_unix = existing.and_then(|mapping| mapping.updated_at_unix);
-		let created_at_unix = existing_created_at_unix.or(observed_at_unix);
-		let updated_at_unix = match (existing_updated_at_unix, observed_at_unix) {
-			(Some(existing), Some(observed)) => Some(existing.max(observed)),
-			(Some(existing), None) => Some(existing),
-			(None, observed) => observed,
-		};
-		let mapping = WorktreeMappingRecord {
-			project_id: project_id.to_owned(),
-			issue_id: issue_id.to_owned(),
-			branch_name: branch_name.to_owned(),
-			worktree_path: PathBuf::from(worktree_path),
-			provenance_source,
-			created_at_unix,
-			updated_at_unix,
-		};
-
-		state.worktrees.insert(issue_id.to_owned(), mapping.clone());
-		state.remember_run_project(project_id, issue_id, None);
-
-		self.upsert_worktree_and_remember_run_project_locked(&mapping)
-	}
-
-	/// Create or replace the retained review handoff projection for one issue lane.
-	pub(crate) fn upsert_review_handoff_marker(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-		marker: &ReviewHandoffMarker,
-	) -> Result<()> {
-		let now = timestamp_parts();
-		let key = ReviewLifecycleKey::new(project_id, issue_id, marker.branch_name());
-		let mut state = self.lock()?;
-		let record = state.review_lifecycle_records.entry(key).or_insert_with(|| {
-			ReviewLifecycleRuntimeRecord {
-				project_id: project_id.to_owned(),
-				issue_id: issue_id.to_owned(),
-				branch_name: marker.branch_name().to_owned(),
-				run_id: marker.run_id().to_owned(),
-				attempt_number: marker.attempt_number(),
-				pr_url: marker.pr_url().to_owned(),
-				target_base_ref_name: marker.target_base_ref_name().map(str::to_owned),
-				pr_head_ref_name: marker.pr_head_ref_name().to_owned(),
-				pr_head_oid: marker.pr_head_oid().to_owned(),
-				head_sha: marker.pr_head_oid().to_owned(),
-				phase: String::from("request_pending"),
-				request_comment_database_id: None,
-				request_created_at_unix_epoch: None,
-				request_description_thumbs_up_count: None,
-				request_retry_count: 0,
-				external_round_count: 0,
-				auto_merge_enabled_at_unix_epoch: None,
-				landing_state: String::from("not_started"),
-				closeout_state: String::from("not_started"),
-				repair_attempt_count: 0,
-				evidence_json: String::from("{}"),
-				next_action: String::new(),
-				updated_at: now.text.clone(),
-				updated_at_unix: now.unix,
-			}
-		});
-		let same_handoff_projection = record.run_id == marker.run_id()
-			&& record.attempt_number == marker.attempt_number()
-			&& record.pr_url == marker.pr_url()
-			&& record.target_base_ref_name.as_deref() == marker.target_base_ref_name()
-			&& record.pr_head_ref_name == marker.pr_head_ref_name()
-			&& record.pr_head_oid == marker.pr_head_oid();
-
-		record.run_id = marker.run_id().to_owned();
-		record.attempt_number = marker.attempt_number();
-		record.pr_url = marker.pr_url().to_owned();
-		record.target_base_ref_name = marker.target_base_ref_name().map(str::to_owned);
-		record.pr_head_ref_name = marker.pr_head_ref_name().to_owned();
-		record.pr_head_oid = marker.pr_head_oid().to_owned();
-
-		if !same_handoff_projection {
-			record.head_sha = marker.pr_head_oid().to_owned();
-			record.phase = String::from("request_pending");
-			record.request_comment_database_id = None;
-			record.request_created_at_unix_epoch = None;
-			record.request_description_thumbs_up_count = None;
-			record.request_retry_count = 0;
-			record.external_round_count = 0;
-			record.auto_merge_enabled_at_unix_epoch = None;
-			record.landing_state = String::from("not_started");
-			record.closeout_state = String::from("not_started");
-			record.repair_attempt_count = 0;
-			record.evidence_json = String::from("{}");
-
-			record.next_action.clear();
-		}
-
-		record.updated_at = now.text;
-		record.updated_at_unix = now.unix;
-
-		self.persist_runtime_state_locked(&state)
-	}
-
-	/// Read the retained review handoff projection for one issue branch.
-	pub(crate) fn review_handoff_marker(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-		branch_name: &str,
-	) -> Result<Option<ReviewHandoffMarker>> {
-		Ok(self.review_lifecycle_record(project_id, issue_id, branch_name)?.map(|record| {
-			ReviewHandoffMarker {
-				run_id: record.run_id().to_owned(),
-				attempt_number: record.attempt_number(),
-				branch_name: record.branch_name().to_owned(),
-				pr_url: record.pr_url().to_owned(),
-				target_base_ref_name: record.target_base_ref_name().map(str::to_owned),
-				pr_head_ref_name: record.pr_head_ref_name().to_owned(),
-				pr_head_oid: record.pr_head_oid().to_owned(),
-			}
-		}))
-	}
-
-	/// Read the runtime-owned review lifecycle record for one retained issue branch.
-	pub(crate) fn review_lifecycle_record(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-		branch_name: &str,
-	) -> Result<Option<ReviewLifecycleRecord>> {
-		let state = self.lock()?;
-		let key = ReviewLifecycleKey::new(project_id, issue_id, branch_name);
-
-		Ok(state.review_lifecycle_records.get(&key).map(ReviewLifecycleRuntimeRecord::as_public))
-	}
-
-	/// Return whether any retained review lifecycle row owns this issue.
-	pub(crate) fn issue_has_review_lifecycle_record(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-	) -> Result<bool> {
-		let state = self.lock()?;
-
-		Ok(state
-			.review_lifecycle_records
-			.values()
-			.any(|record| record.project_id == project_id && record.issue_id == issue_id))
-	}
-
-	/// Create or replace the retained review orchestration projection for one issue lane.
-	pub(crate) fn upsert_review_orchestration_marker(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-		marker: &ReviewOrchestrationMarker,
-	) -> Result<()> {
-		let now = timestamp_parts();
-		let key = ReviewLifecycleKey::new(project_id, issue_id, marker.branch_name());
-		let mut state = self.lock()?;
-		let record = state.review_lifecycle_records.entry(key).or_insert_with(|| {
-			ReviewLifecycleRuntimeRecord {
-				project_id: project_id.to_owned(),
-				issue_id: issue_id.to_owned(),
-				branch_name: marker.branch_name().to_owned(),
-				run_id: marker.run_id().to_owned(),
-				attempt_number: marker.attempt_number(),
-				pr_url: marker.pr_url().to_owned(),
-				target_base_ref_name: None,
-				pr_head_ref_name: marker.branch_name().to_owned(),
-				pr_head_oid: marker.head_sha().to_owned(),
-				head_sha: marker.head_sha().to_owned(),
-				phase: marker.phase().to_owned(),
-				request_comment_database_id: None,
-				request_created_at_unix_epoch: None,
-				request_description_thumbs_up_count: None,
-				request_retry_count: 0,
-				external_round_count: 0,
-				auto_merge_enabled_at_unix_epoch: None,
-				landing_state: String::from("not_started"),
-				closeout_state: String::from("not_started"),
-				repair_attempt_count: 0,
-				evidence_json: String::from("{}"),
-				next_action: String::new(),
-				updated_at: now.text.clone(),
-				updated_at_unix: now.unix,
-			}
-		});
-
-		record.run_id = marker.run_id().to_owned();
-		record.attempt_number = marker.attempt_number();
-		record.pr_url = marker.pr_url().to_owned();
-		record.head_sha = marker.head_sha().to_owned();
-		record.phase = marker.phase().to_owned();
-		record.request_comment_database_id = marker.request_comment_database_id();
-		record.request_created_at_unix_epoch = marker.request_created_at_unix_epoch();
-		record.request_description_thumbs_up_count = marker.request_description_thumbs_up_count();
-		record.request_retry_count = marker.request_retry_count();
-		record.external_round_count = marker.external_round_count();
-		record.auto_merge_enabled_at_unix_epoch = marker.auto_merge_enabled_at_unix_epoch();
-		record.updated_at = now.text;
-		record.updated_at_unix = now.unix;
-
-		self.persist_runtime_state_locked(&state)
-	}
-
-	/// Read retained review orchestration for the current handoff identity.
-	pub(crate) fn review_orchestration_marker(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-		review_handoff: &ReviewHandoffMarker,
-	) -> Result<Option<ReviewOrchestrationMarker>> {
-		let Some(record) =
-			self.review_lifecycle_record(project_id, issue_id, review_handoff.branch_name())?
-		else {
-			return Ok(None);
-		};
-
-		if record.run_id() != review_handoff.run_id()
-			|| record.attempt_number() != review_handoff.attempt_number()
-			|| record.branch_name() != review_handoff.branch_name()
-			|| record.pr_url() != review_handoff.pr_url()
-		{
-			return Ok(None);
-		}
-
-		Ok(Some(ReviewOrchestrationMarker::new(
-			record.run_id().to_owned(),
-			record.attempt_number(),
-			record.branch_name().to_owned(),
-			record.pr_url().to_owned(),
-			record.head_sha().to_owned(),
-			record.phase().to_owned(),
-			record.request_comment_database_id(),
-			record.request_created_at_unix_epoch(),
-			record.request_description_thumbs_up_count(),
-			record.request_retry_count(),
-			record.external_round_count(),
-			record.auto_merge_enabled_at_unix_epoch(),
-		)))
-	}
-
-	/// Create or replace the latest review-policy checkpoint for one run phase.
-	pub(crate) fn upsert_review_policy_checkpoint(
-		&self,
-		input: ReviewPolicyCheckpointInput<'_>,
-	) -> Result<ReviewPolicyCheckpoint> {
-		let now = timestamp_parts();
-		let key = ReviewPolicyKey::new(
-			input.project_id,
-			input.issue_id,
-			input.run_id,
-			input.attempt_number,
-			input.phase,
-		);
-		let record = ReviewPolicyRuntimeRecord {
-			project_id: input.project_id.to_owned(),
-			issue_id: input.issue_id.to_owned(),
-			run_id: input.run_id.to_owned(),
-			attempt_number: input.attempt_number,
-			phase: input.phase.to_owned(),
-			status: input.status.to_owned(),
-			head_sha: input.head_sha.to_owned(),
-			nonclean_rounds: input.nonclean_rounds,
-			details_json: input.details_json.to_owned(),
-			updated_at: now.text.clone(),
-			updated_at_unix: now.unix,
-		};
-		let mut state = self.lock()?;
-
-		state.review_policy_checkpoints.insert(key, record.clone());
-
-		let artifact = evidence_artifact_from_review_checkpoint_input(input, &record, &now)?;
-		let artifact_key = EvidenceArtifactKey::new(
-			&artifact.project_id,
-			&artifact.issue_id,
-			&artifact.artifact_kind,
-			&artifact.key_hash,
-		);
-
-		state.evidence_artifacts.insert(artifact_key, artifact);
-		self.persist_runtime_state_locked(&state)?;
-
-		Ok(record.as_public())
-	}
-
-	/// Read the latest runtime-owned review-policy checkpoint for one run phase.
-	pub(crate) fn review_policy_checkpoint(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-		run_id: &str,
-		attempt_number: i64,
-		phase: &str,
-	) -> Result<Option<ReviewPolicyCheckpoint>> {
-		let state = self.lock()?;
-		let key = ReviewPolicyKey::new(project_id, issue_id, run_id, attempt_number, phase);
-
-		Ok(state.review_policy_checkpoints.get(&key).map(ReviewPolicyRuntimeRecord::as_public))
-	}
-
-	/// Return whether any bounded-review checkpoint row owns this issue.
-	pub(crate) fn issue_has_review_policy_checkpoint(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-	) -> Result<bool> {
-		let state = self.lock()?;
-
-		Ok(state
-			.review_policy_checkpoints
-			.values()
-			.any(|record| record.project_id == project_id && record.issue_id == issue_id))
-	}
-
-	/// Read the latest review checkpoint by its canonical reusable evidence key.
-	pub(crate) fn review_checkpoint_artifact(
-		&self,
-		lookup: ReviewCheckpointArtifactLookup<'_>,
-	) -> Result<Option<ReviewPolicyCheckpoint>> {
-		let state = self.lock()?;
-		let key_json = review_checkpoint_evidence_key_json(
-			lookup.phase,
-			lookup.review_level,
-			lookup.head_sha,
-		)?;
-		let key_hash = evidence_artifact_key_hash("issue_review_checkpoint", &key_json);
-		let key = EvidenceArtifactKey::new(
-			lookup.project_id,
-			lookup.issue_id,
-			"issue_review_checkpoint",
-			&key_hash,
-		);
-
-		state
-			.evidence_artifacts
-			.get(&key)
-			.map(EvidenceArtifactRuntimeRecord::as_review_policy_checkpoint)
-			.transpose()
-	}
-
-	/// Check whether review policy has any non-clean artifact that could fence mutation.
-	pub(crate) fn has_nonclean_review_checkpoint_artifact(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-		phase: &str,
-	) -> Result<bool> {
-		let state = self.lock()?;
-
-		Ok(state.evidence_artifacts.values().any(|record| {
-			record.project_id == project_id
-				&& record.issue_id == issue_id
-				&& record.artifact_kind == "issue_review_checkpoint"
-				&& record.phase == phase
-				&& record.status != "clean"
-		}))
-	}
-
-	/// Clear review-policy checkpoints for one completed run attempt.
-	pub(crate) fn clear_review_policy_checkpoints_for_run_attempt(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-		run_id: &str,
-		attempt_number: i64,
-	) -> Result<()> {
-		let mut state = self.lock()?;
-
-		state.review_policy_checkpoints.retain(|key, _record| {
-			key.project_id != project_id
-				|| key.issue_id != issue_id
-				|| key.run_id != run_id
-				|| key.attempt_number != attempt_number
-		});
-
-		self.delete_review_policy_checkpoints_for_run_attempt_locked(
-			project_id,
-			issue_id,
-			run_id,
-			attempt_number,
-		)
-	}
-
-	/// Record one loop-guardrail observation and return its consecutive count.
-	pub(crate) fn observe_loop_guardrail_checkpoint(
-		&self,
-		input: LoopGuardrailCheckpointInput<'_>,
-	) -> Result<LoopGuardrailCheckpoint> {
-		let now = timestamp_parts();
-		let key = LoopGuardrailKey::new(input.project_id, input.issue_id, input.reason);
-		let mut state = self.lock()?;
-		let previous = state.loop_guardrail_checkpoints.get(&key);
-		let consecutive_count = previous.map_or(1, |record| {
-			if record.fingerprint == input.fingerprint {
-				record.consecutive_count.saturating_add(1)
-			} else {
-				1
-			}
-		});
-		let record = LoopGuardrailRuntimeRecord {
-			project_id: input.project_id.to_owned(),
-			issue_id: input.issue_id.to_owned(),
-			reason: input.reason.to_owned(),
-			fingerprint: input.fingerprint.to_owned(),
-			run_id: input.run_id.to_owned(),
-			attempt_number: input.attempt_number,
-			consecutive_count,
-			details_json: input.details_json.to_owned(),
-			updated_at: now.text,
-			updated_at_unix: now.unix,
-		};
-
-		state.loop_guardrail_checkpoints.insert(key, record.clone());
-		self.persist_runtime_state_locked(&state)?;
-
-		Ok(record.as_public())
-	}
-
-	/// Read one loop-guardrail checkpoint by project, issue, and reason.
-	#[cfg(test)]
-	pub(crate) fn loop_guardrail_checkpoint(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-		reason: &str,
-	) -> Result<Option<LoopGuardrailCheckpoint>> {
-		let state = self.lock()?;
-		let key = LoopGuardrailKey::new(project_id, issue_id, reason);
-
-		Ok(state.loop_guardrail_checkpoints.get(&key).map(LoopGuardrailRuntimeRecord::as_public))
-	}
-
-	/// Clear loop-guardrail checkpoints for one issue.
-	pub(crate) fn clear_loop_guardrail_checkpoints_for_issue(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-	) -> Result<()> {
-		let mut state = self.lock()?;
-
-		state
-			.loop_guardrail_checkpoints
-			.retain(|key, _record| key.project_id != project_id || key.issue_id != issue_id);
-
-		self.delete_loop_guardrail_checkpoints_for_issue_locked(project_id, issue_id)
-	}
-
-	/// Clear one loop-guardrail checkpoint reason for one issue.
-	pub(crate) fn clear_loop_guardrail_checkpoint(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-		reason: &str,
-	) -> Result<()> {
-		let key = LoopGuardrailKey::new(project_id, issue_id, reason);
-		let mut state = self.lock()?;
-
-		state.loop_guardrail_checkpoints.remove(&key);
-
-		self.delete_loop_guardrail_checkpoint_locked(project_id, issue_id, reason)
-	}
-
-	/// Remove the exact review lifecycle record created for one handoff identity.
-	pub(crate) fn clear_review_lifecycle_for_handoff(
-		&self,
-		project_id: &str,
-		issue_id: &str,
-		handoff_marker: &ReviewHandoffMarker,
-		orchestration_marker: &ReviewOrchestrationMarker,
-	) -> Result<()> {
-		let lifecycle_key =
-			ReviewLifecycleKey::new(project_id, issue_id, handoff_marker.branch_name());
-		let mut state = self.lock()?;
-
-		if state
-			.review_lifecycle_records
-			.get(&lifecycle_key)
-			.is_some_and(|record| record.matches_handoff_identity(handoff_marker))
-		{
-			state.review_lifecycle_records.remove(&lifecycle_key);
-		}
-
-		state.review_policy_checkpoints.retain(|key, _record| {
-			key.project_id != project_id
-				|| key.issue_id != issue_id
-				|| key.run_id != orchestration_marker.run_id()
-				|| key.attempt_number != orchestration_marker.attempt_number()
-		});
-		self.persist_runtime_state_locked(&state)?;
-
-		self.delete_review_marker_identity_locked(
-			project_id,
-			issue_id,
-			handoff_marker.branch_name(),
-			orchestration_marker.run_id(),
-			orchestration_marker.attempt_number(),
-		)
-	}
-
-	/// Read the worktree mapping for one issue.
-	pub fn worktree_for_issue(&self, issue_id: &str) -> Result<Option<WorktreeMapping>> {
-		if let Some(sqlite) = &self.sqlite {
-			let sqlite = sqlite.lock().map_err(|_| eyre::eyre!("State store lock poisoned."))?;
-
-			return sqlite
-				.worktree_for_issue(issue_id)
-				.map(|mapping| mapping.map(|mapping| mapping.as_public()));
-		}
-
-		let state = self.lock()?;
-
-		Ok(state.worktrees.get(issue_id).map(WorktreeMappingRecord::as_public))
-	}
-
-	/// List all known worktree mappings.
-	pub fn list_worktrees(&self, project_id: &str) -> Result<Vec<WorktreeMapping>> {
-		let mut state = self.lock_without_refresh()?;
-
-		self.refresh_project_run_metadata_state_locked(&mut state, project_id)?;
-
-		let mut mappings = state
-			.worktrees
-			.values()
-			.filter(|mapping| mapping.project_id == project_id)
-			.map(WorktreeMappingRecord::as_public)
-			.collect::<Vec<_>>();
-
-		mappings.sort_by(|left, right| left.issue_id.cmp(&right.issue_id));
-
-		Ok(mappings)
-	}
-
-	/// Remove the worktree mapping for one issue.
-	pub fn clear_worktree(&self, issue_id: &str) -> Result<()> {
-		let mut state = self.lock()?;
-
-		state.worktrees.remove(issue_id);
-		state.review_lifecycle_records.retain(|key, _record| key.issue_id != issue_id);
-		state.review_policy_checkpoints.retain(|key, _record| key.issue_id != issue_id);
-		self.persist_runtime_state_locked(&state)?;
-
-		self.delete_worktree_and_review_lifecycle_locked(issue_id)
-	}
-
-	/// Remove only the worktree mapping for one issue.
-	pub(crate) fn clear_worktree_mapping(&self, issue_id: &str) -> Result<()> {
-		let mut state = self.lock()?;
-
-		state.worktrees.remove(issue_id);
-		self.persist_runtime_state_locked(&state)?;
-
-		self.delete_worktree_mapping_locked(issue_id)
 	}
 
 	fn lock_without_refresh(&self) -> Result<MutexGuard<'_, StateData>> {
@@ -3903,48 +2528,6 @@ impl StateStore {
 		state.replace_project_run_metadata_state(loaded);
 
 		Ok(())
-	}
-
-	fn refresh_protocol_event_summaries_for_runs_locked(
-		&self,
-		state: &mut StateData,
-		run_ids: &[String],
-	) -> Result<()> {
-		let Some(sqlite) = self.sqlite.as_ref() else {
-			return Ok(());
-		};
-		let sqlite =
-			sqlite.lock().map_err(|_| eyre::eyre!("StateStore SQLite mutex is poisoned."))?;
-
-		sqlite.load_protocol_event_summaries_for_runs(state, run_ids)
-	}
-
-	fn rebuild_protocol_event_summaries_for_runs_locked(
-		&self,
-		state: &mut StateData,
-		run_ids: &[String],
-	) -> Result<()> {
-		let Some(sqlite) = self.sqlite.as_ref() else {
-			return Ok(());
-		};
-		let sqlite =
-			sqlite.lock().map_err(|_| eyre::eyre!("StateStore SQLite mutex is poisoned."))?;
-
-		sqlite.rebuild_protocol_event_summaries_for_runs(state, run_ids)
-	}
-
-	fn upsert_protocol_event_summary_locked(
-		&self,
-		run_id: &str,
-		summary: &ProtocolEventSummaryRecord,
-	) -> Result<()> {
-		let Some(sqlite) = self.sqlite.as_ref() else {
-			return Ok(());
-		};
-		let sqlite =
-			sqlite.lock().map_err(|_| eyre::eyre!("StateStore SQLite mutex is poisoned."))?;
-
-		sqlite.upsert_protocol_event_summary(run_id, summary)
 	}
 
 	fn refresh_run_activity_summaries_for_runs_locked(
@@ -4144,34 +2727,6 @@ impl StateStore {
 			sqlite.lock().map_err(|_| eyre::eyre!("StateStore SQLite mutex is poisoned."))?;
 
 		sqlite.upsert_worktree_and_remember_run_project(mapping)
-	}
-
-	fn append_protocol_event_locked(
-		&self,
-		run_id: &str,
-		event: &ProtocolEventRecord,
-	) -> Result<bool> {
-		let Some(sqlite) = self.sqlite.as_ref() else {
-			return Ok(true);
-		};
-		let sqlite =
-			sqlite.lock().map_err(|_| eyre::eyre!("StateStore SQLite mutex is poisoned."))?;
-
-		sqlite.append_protocol_event(run_id, event)
-	}
-
-	fn protocol_event_locked(
-		&self,
-		run_id: &str,
-		sequence_number: i64,
-	) -> Result<Option<ProtocolEventRecord>> {
-		let Some(sqlite) = self.sqlite.as_ref() else {
-			return Ok(None);
-		};
-		let sqlite =
-			sqlite.lock().map_err(|_| eyre::eyre!("StateStore SQLite mutex is poisoned."))?;
-
-		sqlite.protocol_event(run_id, sequence_number)
 	}
 
 	fn insert_linear_execution_event_if_absent_locked(
@@ -4508,7 +3063,6 @@ fn running_run_attempt_status(status: &str) -> bool {
 	matches!(status, "starting" | "running")
 }
 
-
 #[allow(dead_code)]
 fn validate_decision_contract_record_inputs(
 	project_id: &str,
@@ -4640,193 +3194,4 @@ fn validate_required_execution_program_field(name: &str, value: &str) -> Result<
 	}
 
 	Ok(())
-}
-
-
-fn evidence_artifact_from_review_checkpoint_input(
-	input: ReviewPolicyCheckpointInput<'_>,
-	record: &ReviewPolicyRuntimeRecord,
-	now: &TimestampParts,
-) -> Result<EvidenceArtifactRuntimeRecord> {
-	let key_json =
-		review_checkpoint_evidence_key_json(input.phase, input.review_level, input.head_sha)?;
-	let payload_json = serde_json::json!({
-		"schema": "decodex.review_checkpoint_artifact/1",
-		"phase": input.phase,
-		"review_level": input.review_level,
-		"status": input.status,
-		"head_sha": input.head_sha,
-		"nonclean_rounds": input.nonclean_rounds,
-		"details_json": input.details_json,
-		"source": {
-			"run_id": input.run_id,
-			"attempt_number": input.attempt_number
-		}
-	});
-	let key_hash = evidence_artifact_key_hash("issue_review_checkpoint", &key_json);
-
-	Ok(EvidenceArtifactRuntimeRecord {
-		project_id: record.project_id.clone(),
-		issue_id: record.issue_id.clone(),
-		artifact_kind: String::from("issue_review_checkpoint"),
-		key_hash,
-		phase: record.phase.clone(),
-		status: record.status.clone(),
-		head_sha: Some(record.head_sha.clone()),
-		key_json,
-		payload_json: payload_json.to_string(),
-		source_run_id: record.run_id.clone(),
-		source_attempt_number: record.attempt_number,
-		updated_at: now.text.clone(),
-		updated_at_unix: now.unix,
-	})
-}
-
-fn review_checkpoint_evidence_key_json(
-	phase: &str,
-	review_level: &str,
-	head_sha: &str,
-) -> Result<String> {
-	#[derive(Serialize)]
-	struct ReviewCheckpointEvidenceKey<'a> {
-		schema: &'static str,
-		artifact_kind: &'static str,
-		phase: &'a str,
-		head_sha: &'a str,
-		review_level: &'a str,
-		review_prompt_version: &'static str,
-	}
-
-	serde_json::to_string(&ReviewCheckpointEvidenceKey {
-		schema: "decodex.evidence_key/1",
-		artifact_kind: "issue_review_checkpoint",
-		phase,
-		head_sha,
-		review_level,
-		review_prompt_version: REVIEW_CHECKPOINT_PROMPT_VERSION,
-	})
-	.map_err(|error| eyre::eyre!("failed to serialize review checkpoint evidence key: {error}"))
-}
-
-fn evidence_artifact_key_hash(artifact_kind: &str, key_json: &str) -> String {
-	let payload = format!("{artifact_kind}\n{key_json}");
-	let digest = Sha256::digest(payload.as_bytes());
-	let mut hash = String::with_capacity(64);
-
-	for byte in digest {
-		hash.push(char::from(b"0123456789abcdef"[(byte >> 4) as usize]));
-		hash.push(char::from(b"0123456789abcdef"[(byte & 0x0f) as usize]));
-	}
-
-	hash
-}
-
-fn protocol_event_payload_sha256(payload: &str) -> String {
-	let digest = Sha256::digest(payload.as_bytes());
-	let mut hash = String::with_capacity(64);
-
-	for byte in digest {
-		hash.push(char::from(b"0123456789abcdef"[(byte >> 4) as usize]));
-		hash.push(char::from(b"0123456789abcdef"[(byte & 0x0f) as usize]));
-	}
-
-	hash
-}
-
-fn protocol_event_is_terminal_thread_archive(event_type: &str) -> bool {
-	TERMINAL_THREAD_ARCHIVE_EVENT_TYPES.contains(&event_type)
-}
-
-fn protocol_event_can_be_discarded_after_archive(event: &ProtocolEventRecord) -> bool {
-	!protocol_event_is_terminal_thread_archive(&event.event_type)
-		&& event.event_type != DISCARDED_POST_ARCHIVE_PROTOCOL_EVENT_TYPE
-}
-
-fn protocol_event_conflict_should_be_discarded_after_archive(
-	existing: &ProtocolEventRecord,
-	candidate: &ProtocolEventRecord,
-) -> bool {
-	protocol_event_is_terminal_thread_archive(&existing.event_type)
-		&& protocol_event_can_be_discarded_after_archive(candidate)
-}
-
-fn protocol_event_is_discarded_post_archive_collision(
-	existing: &ProtocolEventRecord,
-	candidate: &ProtocolEventRecord,
-) -> bool {
-	existing.event_type == DISCARDED_POST_ARCHIVE_PROTOCOL_EVENT_TYPE
-		&& candidate.event_type == DISCARDED_POST_ARCHIVE_PROTOCOL_EVENT_TYPE
-		&& !existing.is_idempotent_replay_of(candidate)
-}
-
-fn discarded_post_archive_protocol_event(mut event: ProtocolEventRecord) -> ProtocolEventRecord {
-	if event.event_type == DISCARDED_POST_ARCHIVE_PROTOCOL_EVENT_TYPE {
-		return event;
-	}
-
-	event.sequence_number = discarded_post_archive_protocol_sequence(&event);
-	event.event_type = DISCARDED_POST_ARCHIVE_PROTOCOL_EVENT_TYPE.to_owned();
-
-	event
-}
-
-fn discarded_post_archive_protocol_event_with_log(
-	run_id: &str,
-	event: ProtocolEventRecord,
-) -> ProtocolEventRecord {
-	let original_sequence_number = event.sequence_number;
-	let original_event_type = event.event_type.clone();
-	let discarded = discarded_post_archive_protocol_event(event);
-
-	tracing::info!(
-		run_id,
-		original_sequence_number,
-		original_event_type,
-		discarded_sequence_number = discarded.sequence_number,
-		discarded_event_type = discarded.event_type.as_str(),
-		"Discarded late app-server protocol event after terminal thread archive barrier; child protocol activity is isolated from parent journal and closeout state."
-	);
-
-	discarded
-}
-
-fn discarded_post_archive_protocol_sequence(event: &ProtocolEventRecord) -> i64 {
-	let payload =
-		format!("{}\n{}\n{}", event.sequence_number, event.event_type, event.payload_sha256);
-	let digest = Sha256::digest(payload.as_bytes());
-	let mut bytes = [0_u8; 8];
-
-	bytes.copy_from_slice(&digest[..8]);
-
-	let slot = i64::from_be_bytes(bytes) & i64::MAX;
-
-	if slot == i64::MAX { i64::MIN } else { -1 - slot }
-}
-
-fn next_discarded_post_archive_sequence_after_collision(sequence_number: i64) -> Result<i64> {
-	if sequence_number == i64::MIN {
-		eyre::bail!("Post-archive discarded protocol event sequence space is exhausted.");
-	}
-
-	Ok(sequence_number - 1)
-}
-
-fn ensure_protocol_event_replay_matches(
-	run_id: &str,
-	existing: &ProtocolEventRecord,
-	candidate: &ProtocolEventRecord,
-) -> Result<()> {
-	if existing.is_idempotent_replay_of(candidate) {
-		return Ok(());
-	}
-
-	eyre::bail!(
-		"Protocol event `{run_id}` sequence `{}` conflicts with an existing runtime journal event: \
-		 existing event_type=`{}` payload_sha256=`{}`, candidate event_type=`{}` payload_sha256=`{}`.",
-		candidate.sequence_number,
-		existing.event_type,
-		existing.payload_sha256,
-		candidate.event_type,
-		candidate.payload_sha256,
-	);
 }
